feat(create event; #25 US 02.01.01): complete model definition and bind
There are still some issues to resolve. Currently the implementation
uses a `FragmentContainerView`, but it may be the case that another
class can be used. `ViewSwitcher` doesn't work for the use case since it
only allows switching between 2 child views.

The next commit will be an attempt at using `FrameLayoutView` and swapping
the input section manually which seems to be the way the implementation
is headed. If that doesn't work it will changed back to
`FragmentContainerView`.